requirements.txt
fastapi 
uvicorn 
pymongo 
redis 
python-dotenv 
requests 
pytest
motor
python-dotenv
beautifulsoup4
google-search-results
newsapi-python
scikit-learn
spacy
torch
transformers
accelerate
pymongo
streamlit